add role based auth

+++ users/src/main/java/com/keskin/users/service/impl/UserServiceImpl.java
@@ -1,5 +1,6 @@
 package com.keskin.users.service.impl;
 
+import com.keskin.users.config.UserContextHolder;
 import com.keskin.users.dto.request.UpdateUserRequestDto;
 import com.keskin.users.dto.UserDto;
 import com.keskin.users.entity.User;
@@ -7,6 +8,7 @@ import com.keskin.users.mapper.UserMapper;
 import com.keskin.users.repository.UserRepository;
 import com.keskin.users.service.IUserService;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -20,43 +22,68 @@ public class UserServiceImpl implements IUserService {
     private final UserRepository userRepository;
     private final UserMapper userMapper;
 
-    private User findUser(Long id){
+
+    private User findUser(Long id) {
         return userRepository.findById(id).orElseThrow(() ->
-                new IllegalArgumentException("User with the id not found : " + id));
+                new IllegalArgumentException("User with the id not found: " + id));
     }
+
+    /**
+     * Private method for centralized managing
+     */
+    private void validateOwnershipOrAdmin(User user) {
+        boolean isOwner = user.getEmail().equals(UserContextHolder.getEmail());
+        boolean isAdmin = UserContextHolder.isAdmin();
+
+        if (!isOwner && !isAdmin) {
+            throw new AccessDeniedException("You don't have permission for this action!");
+        }
+    }
+
 
     @Override
     public UserDto getUser(Long id) {
         User user = findUser(id);
+        validateOwnershipOrAdmin(user);
         return userMapper.entityToDto(user);
     }
 
     @Override
     public List<UserDto> getAllUsers() {
+        if (!UserContextHolder.isAdmin()) {
+            throw new AccessDeniedException("Only admins can list all users!");
+        }
         return userRepository.findAll()
                 .stream()
-                .map(user -> userMapper.entityToDto(user))
+                .map(userMapper::entityToDto)
                 .toList();
     }
 
     @Override
     public UserDto updateUser(Long id, UpdateUserRequestDto requestDto) {
-        //add existing mail check when free
         User user = findUser(id);
-        User updatedUser = userMapper.updateRequestToEntity(requestDto, user);
-        userRepository.save(updatedUser);
-        return userMapper.entityToDto(updatedUser);
+        validateOwnershipOrAdmin(user);
+
+        //prevent email from overriding
+        String originalEmail = user.getEmail();
+        userMapper.updateRequestToEntity(requestDto, user);
+        user.setEmail(originalEmail);
+
+        return userMapper.entityToDto(userRepository.save(user));
     }
 
     @Override
     public void deleteUser(Long id) {
         User user = findUser(id);
+        validateOwnershipOrAdmin(user);
         userRepository.delete(user);
     }
 
     @Override
     public void toggleActive(Long id) {
         User user = findUser(id);
+        validateOwnershipOrAdmin(user);
+
         user.setActive(!user.isActive());
         userRepository.save(user);
     }
